Add some nice timing analysis reporting comparing Merge Sort and Quick Sort with standard deviation and mean; eliminate threads argument; optimize Merge Sort with alloca (note that this may create issues with huge data sets).

+++ src/sortbench.cc
@@ -16,10 +16,18 @@
 // along with sortbench.  If not, see <https://www.gnu.org/licenses/>.
 //
 
+// C headers
 #include <stdio.h>
 #include <math.h>
 
+// C++ headers
+#include <iostream>
 #include <map>
+#include <vector>
+#include <chrono>
+
+// Project-specific
+#include "common.h"
 #include "algo.h"
 #include "merge_sort.h"
 #include "quick_sort.h"
@@ -33,7 +41,7 @@ void PrintUsage()
 {
   printf("sortbench\n" );
   printf("Usage:\n" );
-  printf("\tsortbench <array_size> <array_total> <thread_total>\n");
+  printf("\tsortbench <array_size> <iteration_total>\n");
 }
 
 // printArray
@@ -51,13 +59,10 @@ void PrintArray(const hedger::S_T *array, size_t n)
   printf("\n");
 }
 
-// createDataSet
-//
-// Allocates and fills an array with unique pseudo-random values.
-//
-// Entry: size
-// Exit:  pointer to array
-hedger::S_T * CreateUniqueDataSet(size_t size)
+// AllocArray
+// Entry: size of array in elements
+// Exit:  pointer to array, or nullptr == error
+hedger::S_T * AllocArray(size_t size)
 {
   // Allocate the array
   hedger::S_T *array = new hedger::S_T[ size ];
@@ -66,7 +71,26 @@ hedger::S_T * CreateUniqueDataSet(size_t size)
     printf(" %s: Allocation error", __FUNCTION__);
     return nullptr;
   }
+  return array;
+}
 
+// FreeArray
+// Entry: pointer to array
+void FreeArray(hedger::S_T *array)
+{
+  if (array) {
+    delete array;
+  }
+}
+
+// createDataSet
+//
+// Allocates and fills an array with unique pseudo-random values.
+//
+// Entry: size
+// Exit:  pointer to array
+hedger::S_T * CreateUniqueDataSet(hedger::S_T *array, size_t size)
+{
   // In-situ STL map<> for quick lookup of already-used random values
   std::map<size_t, hedger::S_T> occupancy;
 
@@ -109,42 +133,76 @@ void TestBtree(size_t array_size)
 {
   hedger::ScapegoatTree btree;
 
-  hedger::S_T *array = CreateUniqueDataSet( array_size );
+  hedger::S_T *array = AllocArray (array_size);
+  if (array) {
+    CreateUniqueDataSet( array, array_size );
 
-  if (!array) {
-    // TODO: LOG ERROR
-    return;
+    if (!array) {
+      // TODO: LOG ERROR
+      return;
+    }
+
+    for (size_t i = 0; i < array_size; i++) {
+      btree.Add(array[i]);
+    }
+
+    btree.Print();
+    int maxDepth = btree.MaxDepth();
+    printf( "\nMAX DEPTH: %d\n", maxDepth);
+
+    for (hedger::S_T i = 0; i < (hedger::S_T) array_size; i++)
+    {
+      hedger::Node *node = btree.Find(i);
+      printf("FIND: %08x\t", (int) (*(int *)node) );
+    }
+
+    printf("\n");
+    btree.DeleteKey(17);
+
+    for (hedger::S_T i = 0; i < (hedger::S_T) array_size; i++)
+    {
+      hedger::Node *node = btree.Find(i);
+      printf("FIND: %08x\t", (int) (*(int *)node) );
+    }
+
+    printf("\n");
+    FreeArray(array);
+  } else {
+    printf("%s:%d Error allocating array.\n", __FUNCTION__, __LINE__);
   }
-
-  for (size_t i = 0; i < array_size; i++) {
-    btree.Add(array[i]);
-  }
-
-  btree.Print();
-  int maxDepth = btree.MaxDepth();
-  printf( "\nMAX DEPTH: %d\n", maxDepth);
-
-  for (hedger::S_T i = 0; i < (hedger::S_T) array_size; i++)
-  {
-    hedger::Node *node = btree.Find(i);
-    printf("FIND: %08x\t", (int) (*(int *)node) );
-  }
-
-  printf("\n");
-  btree.DeleteKey(17);
-
-  for (hedger::S_T i = 0; i < (hedger::S_T) array_size; i++)
-  {
-    hedger::Node *node = btree.Find(i);
-    printf("FIND: %08x\t", (int) (*(int *)node) );
-  }
-
-  printf("\n");
-
-  delete array;
   return;
 }
 
+// ReportTiming
+// Calculate and report mean and standard deviation of timing.
+// Entry: vector of times
+//        name of algorithm
+// TODO: Move this to its own utility class
+void ReportTiming(std::vector<double>& v, int iteration_tot, const char *name)
+{
+    // Calculate average (mu)
+    double mu, sigma;
+    double accum = 0.0;
+    for (auto i : v) {
+      accum = accum + i;
+    }
+    mu = (double) accum / (double) iteration_tot;
+
+    // Calculate std deviation (sigma)
+    accum = 0.0;
+    for (auto i : v) {
+      accum += pow(( (double) i - mu), 2);
+    }
+    sigma = sqrt(accum / iteration_tot);
+
+    // Print report
+        std::cout << COUT_YELLOW << name << ":" << COUT_NORMAL << std::endl;
+    std::cout << "TRIES TOT: " << iteration_tot << std::endl;
+    std::cout << "TIME MU: " << mu << std::endl;
+    std::cout << "TIME SIGMA: " << sigma << std::endl;
+}
+
+// main
 int main(int argc, const char **argv)
 {
   int result = 0;
@@ -156,12 +214,8 @@ int main(int argc, const char **argv)
     printUsage();
     return -1;
   }
-
-
   sscanf(argv[1], "%d", (int *) &array_size);
-
   TestBtree(array_size);
-
 #else
 
   // Seed random number generator (use seconds since epoch)
@@ -171,37 +225,62 @@ int main(int argc, const char **argv)
   hedger::MergeSort *mergeSort = new hedger::MergeSort();
   hedger::QuickSort *quickSort = new hedger::QuickSort();
 
-  // Gather user parameters: array_size, array_tot, thread_tot
+  // Gather user parameters: array_size, iteration_tot
   size_t array_size;
-  int array_tot, thread_tot;
-  array_size = array_tot = thread_tot = 0;
+  int iteration_tot;
+  array_size = iteration_tot = 0;
 
-  if (argc < 4) {
+  if (argc < 3) {
     PrintUsage();
     return -1;
   }
   sscanf(argv[1], "%d", (int *) &array_size);
-  sscanf(argv[2], "%d", &array_tot);
-  sscanf(argv[3], "%d", &thread_tot);
+  sscanf(argv[2], "%d", &iteration_tot);
 
   // Validate params
-  if (!array_size || !array_tot || !thread_tot) {
+  if (!array_size || !iteration_tot) {
     PrintUsage();
     return -1;
   }
 
   // Create the unique unsorted data set
-  hedger::S_T *array = CreateUniqueDataSet( array_size );
 
+  // Timing variables for statistical analysis
+  std::vector<double> time_a, time_b;
+  int iteration_count = iteration_tot;
+
+  // Allocate our array
+  hedger::S_T *array = AllocArray(array_size);
   if (array) {
-    printf( "BEFORE:\n" );
-    PrintArray(array, array_size);
+    do {
+      using FpMilliseconds =
+        std::chrono::duration<float, std::chrono::milliseconds::period>;
 
-    //Test( mergeSort, array, array_size );
-    Test( quickSort, array, array_size );
+      // Inline timing profiling
+      // TODO: Improvement: Should have a data structure (map of vectors?) capable of storing
+      // all the algorithm rather than time_a and time_b.
+      // Merge sort
+      array = CreateUniqueDataSet(array, array_size);                 // randomise data
+      auto start = std::chrono::high_resolution_clock::now();    // mark start time
+      Test( mergeSort, array, array_size );                      // Do work
+      auto stop = std::chrono::high_resolution_clock::now();     // mark end time
+      auto ms = FpMilliseconds(stop - start);                    // get elapsed time in ms
+      double ms_float = ms.count();                              // get as a float
+      time_a.push_back(ms_float);                                // save in our timing array
 
-    printf( "AFTER:\n" );
-    PrintArray(array, array_size);
+      // Quick sort
+      array = CreateUniqueDataSet(array, array_size);
+      start = std::chrono::high_resolution_clock::now();
+      Test( quickSort, array, array_size );
+      stop = std::chrono::high_resolution_clock::now();
+      ms = FpMilliseconds(stop - start);
+      ms_float = ms.count();
+      time_b.push_back(ms_float);
+    } while (--iteration_count);
+
+    // Calculate mean (mu)
+    ReportTiming(time_a, iteration_tot, "Merge Sort");
+    ReportTiming(time_b, iteration_tot, "Quick Sort");
   } else {
     // TODO: SEND TO LOGGER
     printf("Failed to allocate data set array.\n");
